+++ sdk/go/aws/lambda/layerVersion.go
@@ -15,6 +15,7 @@ import (
 //
 // For information about Lambda Layers and how to use them, see [AWS Lambda Layers](https://docs.aws.amazon.com/lambda/latest/dg/configuration-layers.html)
 //
+// ## Example Usage
 // ## Specifying the Deployment Package
 //
 // AWS Lambda Layers expect source code to be provided as a deployment package whose structure varies depending on which `compatibleRuntimes` this layer specifies.
@@ -45,6 +46,8 @@ type LayerVersion struct {
 	Arn pulumi.StringOutput `pulumi:"arn"`
 	// The path to the function's deployment package within the local filesystem. If defined, The `s3_`-prefixed options cannot be used.
 	Code pulumi.ArchiveOutput `pulumi:"code"`
+	// The compatible architectures for the specific Lambda Layer Version. Valid values are `arm64` and `x8664`. If not supplied, a value of `null` is assumed and will default to the value used for the function architecture.
+	CompatibleArchitectures pulumi.StringArrayOutput `pulumi:"compatibleArchitectures"`
 	// A list of [Runtimes](https://docs.aws.amazon.com/lambda/latest/dg/API_PublishLayerVersion.html#SSS-PublishLayerVersion-request-CompatibleRuntimes) this layer is compatible with. Up to 5 runtimes can be specified.
 	CompatibleRuntimes pulumi.StringArrayOutput `pulumi:"compatibleRuntimes"`
 	// The date this resource was created.
@@ -111,6 +114,8 @@ type layerVersionState struct {
 	Arn *string `pulumi:"arn"`
 	// The path to the function's deployment package within the local filesystem. If defined, The `s3_`-prefixed options cannot be used.
 	Code pulumi.Archive `pulumi:"code"`
+	// The compatible architectures for the specific Lambda Layer Version. Valid values are `arm64` and `x8664`. If not supplied, a value of `null` is assumed and will default to the value used for the function architecture.
+	CompatibleArchitectures []string `pulumi:"compatibleArchitectures"`
 	// A list of [Runtimes](https://docs.aws.amazon.com/lambda/latest/dg/API_PublishLayerVersion.html#SSS-PublishLayerVersion-request-CompatibleRuntimes) this layer is compatible with. Up to 5 runtimes can be specified.
 	CompatibleRuntimes []string `pulumi:"compatibleRuntimes"`
 	// The date this resource was created.
@@ -146,6 +151,8 @@ type LayerVersionState struct {
 	Arn pulumi.StringPtrInput
 	// The path to the function's deployment package within the local filesystem. If defined, The `s3_`-prefixed options cannot be used.
 	Code pulumi.ArchiveInput
+	// The compatible architectures for the specific Lambda Layer Version. Valid values are `arm64` and `x8664`. If not supplied, a value of `null` is assumed and will default to the value used for the function architecture.
+	CompatibleArchitectures pulumi.StringArrayInput
 	// A list of [Runtimes](https://docs.aws.amazon.com/lambda/latest/dg/API_PublishLayerVersion.html#SSS-PublishLayerVersion-request-CompatibleRuntimes) this layer is compatible with. Up to 5 runtimes can be specified.
 	CompatibleRuntimes pulumi.StringArrayInput
 	// The date this resource was created.
@@ -183,6 +190,8 @@ func (LayerVersionState) ElementType() reflect.Type {
 type layerVersionArgs struct {
 	// The path to the function's deployment package within the local filesystem. If defined, The `s3_`-prefixed options cannot be used.
 	Code pulumi.Archive `pulumi:"code"`
+	// The compatible architectures for the specific Lambda Layer Version. Valid values are `arm64` and `x8664`. If not supplied, a value of `null` is assumed and will default to the value used for the function architecture.
+	CompatibleArchitectures []string `pulumi:"compatibleArchitectures"`
 	// A list of [Runtimes](https://docs.aws.amazon.com/lambda/latest/dg/API_PublishLayerVersion.html#SSS-PublishLayerVersion-request-CompatibleRuntimes) this layer is compatible with. Up to 5 runtimes can be specified.
 	CompatibleRuntimes []string `pulumi:"compatibleRuntimes"`
 	// Description of what your Lambda Layer does.
@@ -205,6 +214,8 @@ type layerVersionArgs struct {
 type LayerVersionArgs struct {
 	// The path to the function's deployment package within the local filesystem. If defined, The `s3_`-prefixed options cannot be used.
 	Code pulumi.ArchiveInput
+	// The compatible architectures for the specific Lambda Layer Version. Valid values are `arm64` and `x8664`. If not supplied, a value of `null` is assumed and will default to the value used for the function architecture.
+	CompatibleArchitectures pulumi.StringArrayInput
 	// A list of [Runtimes](https://docs.aws.amazon.com/lambda/latest/dg/API_PublishLayerVersion.html#SSS-PublishLayerVersion-request-CompatibleRuntimes) this layer is compatible with. Up to 5 runtimes can be specified.
 	CompatibleRuntimes pulumi.StringArrayInput
 	// Description of what your Lambda Layer does.